+++ client/jetbrains/src/main/java/com/sourcegraph/cody/recipes/RunCodyAppAction.java
@@ -3,9 +3,9 @@ package com.sourcegraph.cody.recipes;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.project.DumbAwareAction;
 import com.intellij.openapi.project.Project;
+import com.sourcegraph.cody.config.AccountType;
+import com.sourcegraph.cody.config.CodyAuthenticationManager;
 import com.sourcegraph.cody.localapp.LocalAppManager;
-import com.sourcegraph.config.ConfigUtil;
-import com.sourcegraph.config.SettingsComponent;
 import org.jetbrains.annotations.NotNull;
 
 public class RunCodyAppAction extends DumbAwareAction {
@@ -22,7 +22,8 @@ public class RunCodyAppAction extends DumbAwareAction {
       return;
     }
     if (LocalAppManager.isPlatformSupported()
-        && ConfigUtil.getInstanceType(project) == SettingsComponent.InstanceType.LOCAL_APP) {
+        && CodyAuthenticationManager.getInstance().getDefaultAccountType(project)
+            == AccountType.LOCAL_APP) {
       if (LocalAppManager.isLocalAppInstalled() && !LocalAppManager.isLocalAppRunning()) {
         showAction(e);
         return;