+++ client/jetbrains/src/main/java/com/sourcegraph/telemetry/GraphQlLogger.java
@@ -8,6 +8,8 @@ import com.sourcegraph.cody.PluginUtil;
 import com.sourcegraph.cody.agent.CodyAgent;
 import com.sourcegraph.cody.agent.protocol.CompletionEvent;
 import com.sourcegraph.cody.agent.protocol.Event;
+import com.sourcegraph.cody.config.CodyApplicationSettings;
+import com.sourcegraph.cody.config.SourcegraphServerPath;
 import com.sourcegraph.config.ConfigUtil;
 import java.util.concurrent.CompletableFuture;
 import org.jetbrains.annotations.NotNull;
@@ -17,16 +19,19 @@ public class GraphQlLogger {
   private static final Gson gson = new GsonBuilder().serializeNulls().create();
 
   public static CompletableFuture<Boolean> logInstallEvent(@NotNull Project project) {
-    if (ConfigUtil.getAnonymousUserId() != null && project.isDisposed()) {
-      var event = createEvent(project, "CodyInstalled", new JsonObject());
+    CodyApplicationSettings codyApplicationSettings = CodyApplicationSettings.getInstance();
+    if (codyApplicationSettings.getAnonymousUserId() != null && project.isDisposed()) {
+      var event = createEvent(ConfigUtil.getServerPath(project), "CodyInstalled", new JsonObject());
       return logEvent(project, event);
     }
     return CompletableFuture.completedFuture(false);
   }
 
   public static void logUninstallEvent(@NotNull Project project) {
-    if (ConfigUtil.getAnonymousUserId() != null) {
-      Event event = createEvent(project, "CodyUninstalled", new JsonObject());
+    CodyApplicationSettings codyApplicationSettings = CodyApplicationSettings.getInstance();
+    if (codyApplicationSettings.getAnonymousUserId() != null) {
+      Event event =
+          createEvent(ConfigUtil.getServerPath(project), "CodyUninstalled", new JsonObject());
       logEvent(project, event);
     }
   }
@@ -42,14 +47,15 @@ public class GraphQlLogger {
     eventParameters.addProperty("displayDuration", displayDurationMs);
     eventParameters.addProperty("isAnyKnownPluginEnabled", PluginUtil.isAnyKnownPluginEnabled());
     JsonObject updatedEventParameters = addCompletionEventParams(eventParameters, params);
-    logEvent(project, createEvent(project, eventName, updatedEventParameters));
+    logEvent(
+        project, createEvent(ConfigUtil.getServerPath(project), eventName, updatedEventParameters));
   }
 
   public static void logAutocompleteAcceptedEvent(
       @NotNull Project project, @Nullable CompletionEvent.Params params) {
     String eventName = "CodyJetBrainsPlugin:completion:accepted";
     JsonObject eventParameters = addCompletionEventParams(new JsonObject(), params);
-    logEvent(project, createEvent(project, eventName, eventParameters));
+    logEvent(project, createEvent(ConfigUtil.getServerPath(project), eventName, eventParameters));
   }
 
   private static JsonObject addCompletionEventParams(
@@ -73,24 +79,27 @@ public class GraphQlLogger {
   public static void logCodyEvent(
       @NotNull Project project, @NotNull String componentName, @NotNull String action) {
     var eventName = "CodyJetBrainsPlugin:" + componentName + ":" + action;
-    logEvent(project, createEvent(project, eventName, new JsonObject()));
+    logEvent(project, createEvent(ConfigUtil.getServerPath(project), eventName, new JsonObject()));
   }
 
   @NotNull
   private static Event createEvent(
-      @NotNull Project project, @NotNull String eventName, @NotNull JsonObject eventParameters) {
-    var updatedEventParameters = addGlobalEventParameters(eventParameters, project);
-    String anonymousUserId = ConfigUtil.getAnonymousUserId();
+      @NotNull SourcegraphServerPath sourcegraphServerPath,
+      @NotNull String eventName,
+      @NotNull JsonObject eventParameters) {
+    var updatedEventParameters = addGlobalEventParameters(eventParameters, sourcegraphServerPath);
+    CodyApplicationSettings codyApplicationSettings = CodyApplicationSettings.getInstance();
+    String anonymousUserId = codyApplicationSettings.getAnonymousUserId();
     return new Event(
         eventName, anonymousUserId != null ? anonymousUserId : "", "", updatedEventParameters);
   }
 
   @NotNull
   private static JsonObject addGlobalEventParameters(
-      @NotNull JsonObject eventParameters, @NotNull Project project) {
+      @NotNull JsonObject eventParameters, @NotNull SourcegraphServerPath sourcegraphServerPath) {
     // project specific properties
     var updatedEventParameters = eventParameters.deepCopy();
-    updatedEventParameters.addProperty("serverEndpoint", ConfigUtil.getSourcegraphUrl(project));
+    updatedEventParameters.addProperty("serverEndpoint", sourcegraphServerPath.getUrl());
     // Extension specific properties
     JsonObject extensionDetails = new JsonObject();
     extensionDetails.addProperty("ide", "JetBrains");