+++ client/jetbrains/src/main/java/com/sourcegraph/vcs/RepoUtil.java
@@ -6,8 +6,8 @@ import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.vcsUtil.VcsUtil;
+import com.sourcegraph.cody.config.CodyProjectSettings;
 import com.sourcegraph.common.ErrorNotification;
-import com.sourcegraph.config.ConfigUtil;
 import git4idea.GitVcs;
 import git4idea.repo.GitRepository;
 import java.io.File;
@@ -31,6 +31,7 @@ public class RepoUtil {
     String relativePath = "";
     String remoteUrl = "";
     String remoteBranchName = "";
+    CodyProjectSettings codyProjectSettings = CodyProjectSettings.getInstance(project);
     try {
       String repoRootPath = getRepoRootPath(project, file);
       if (repoRootPath == null) {
@@ -47,14 +48,14 @@ public class RepoUtil {
       }
 
       remoteUrl = getRemoteRepoUrl(project, file);
-      remoteUrl = doReplacements(project, remoteUrl);
+      remoteUrl = doReplacements(codyProjectSettings, remoteUrl);
 
       // If the current branch doesn't exist on the remote or if the remote
       // for the current branch doesn't correspond with the sourcegraph remote,
       // use the default branch for the project.
       remoteBranchName = getRemoteBranchName(project, file);
       if (remoteBranchName == null || !remoteUrl.contains(remoteBranchName)) {
-        remoteBranchName = ConfigUtil.getDefaultBranchName(project);
+        remoteBranchName = codyProjectSettings.getDefaultBranchName();
       }
     } catch (Exception err) {
       String message;
@@ -70,7 +71,7 @@ public class RepoUtil {
     return new RepoInfo(
         vcsType,
         remoteUrl,
-        remoteBranchName != null ? remoteBranchName : ConfigUtil.getDefaultBranchName(project),
+        remoteBranchName != null ? remoteBranchName : codyProjectSettings.getDefaultBranchName(),
         relativePath);
   }
 
@@ -101,9 +102,10 @@ public class RepoUtil {
     return repository.getRoot().getName();
   }
 
-  private static String doReplacements(@NotNull Project project, @NotNull String remoteUrl) {
+  private static String doReplacements(
+      @NotNull CodyProjectSettings codyProjectSettings, @NotNull String remoteUrl) {
     String remoteUrlWithReplacements = remoteUrl;
-    String r = ConfigUtil.getRemoteUrlReplacements(project);
+    String r = codyProjectSettings.getRemoteUrlReplacements();
     String[] replacements = r.trim().split("\\s*,\\s*");
     if (replacements.length % 2 == 0) {
       for (int i = 0; i < replacements.length; i += 2) {