+++ client/jetbrains/src/main/java/com/sourcegraph/website/FileActionBase.java
@@ -9,9 +9,9 @@ import com.intellij.openapi.fileEditor.FileEditorManager;
 import com.intellij.openapi.project.DumbAwareAction;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
+import com.sourcegraph.cody.config.AccountType;
+import com.sourcegraph.cody.config.CodyAuthenticationManager;
 import com.sourcegraph.common.ErrorNotification;
-import com.sourcegraph.config.ConfigUtil;
-import com.sourcegraph.config.SettingsComponent;
 import com.sourcegraph.find.PreviewContent;
 import com.sourcegraph.find.SourcegraphVirtualFile;
 import com.sourcegraph.vcs.RepoInfo;
@@ -138,6 +138,7 @@ public abstract class FileActionBase extends DumbAwareAction {
     }
     e.getPresentation()
         .setEnabled(
-            ConfigUtil.getInstanceType(project) != SettingsComponent.InstanceType.LOCAL_APP);
+            CodyAuthenticationManager.getInstance().getDefaultAccountType(project)
+                != AccountType.LOCAL_APP);
   }
 }