+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/api/SourcegraphAuthenticationException.kt
@@ -1,0 +1,5 @@
+package com.sourcegraph.cody.api
+
+import java.io.IOException
+
+class SourcegraphAuthenticationException(message: String) : IOException(message)