+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/api/SourcegraphStatusCodeException.kt
@@ -1,0 +1,4 @@
+package com.sourcegraph.cody.api
+
+class SourcegraphStatusCodeException(message: String?, val statusCode: Int) :
+    SourcegraphConfusingException(message)