+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/config/CodyAccountDetails.kt
@@ -1,0 +1,8 @@
+package com.sourcegraph.cody.config
+
+import com.intellij.collaboration.auth.AccountDetails
+
+class CodyAccountDetails(val username: String, val displayName: String?, val avatarURL: String?) : AccountDetails {
+  override val name: String
+    get() = displayName ?: username
+}