+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/config/SourcegraphParseException.kt
@@ -1,0 +1,3 @@
+package com.sourcegraph.cody.config
+
+class SourcegraphParseException(message: String) : RuntimeException(message)